backend/requirements.txt
fastapi
uvicorn
pydantic
beautifulsoup4
pandas
numpy
openpyxl
python-docx
pypdf
pdfplumber
Pillow
requests
jsonschema
networkx
lizard
tree-sitter
tree-sitter-c
clang
reportlab
tiktoken
google-genai
google-generativeai
psycopg2-binary
python-multipart
libclang
httpx
